Refactor: organize game use cases into subpackages
This commit reorganizes the use case classes within the `domain/game` module into more specific subpackages: `game`, `hint`, `input`, `time`, and `visuals`.

**Key Changes:**

-   **Package Restructuring:**
    -   `AutoClearNotesUseCase`, `ClearActiveGameUseCase`, `GetGameUseCase`, `ResetGameUseCase`, and `SaveGameUseCase` have been moved to the `com.example.domain.game.usecases.game` package.
    -   `FocusOnHintCellsUseCase`, `GenerateHintLogUseCase`, `ProvideHintUseCase`, `RevealHintOnGridUseCase`, and `RevealLastHintLogUseCase` have been moved to the `com.example.domain.game.usecases.hint` package.
    -   `InputNumberUseCase`, `RedoOperationUseCase`, `SelectCellUseCase`, and `UndoOperationUseCase` have been moved to the `com.example.domain.game.usecases.input` package.
    -   `GetBestTimeUseCase`, `GetElapsedTimeUseCase`, and `SaveElapsedTimeUseCase` have been moved to the `com.example.domain.game.usecases.time` package.
    -   `ClearHighlightedRowAndColumnUseCase`, `HighlightMatchingNumbersUseCase`, `HighlightRowAndColumnUseCase`, and the new `MarkRuleBreakingCellsUseCase` have been moved to or created in the `com.example.domain.game.usecases.visuals` package.
-   **New Use Case:**
    -   Introduced `MarkRuleBreakingCellsUseCase` in `com.example.domain.game.usecases.visuals` which uses `sudokuGrid.markRuleBreakingCells()` to update the grid.
-   **Import Updates:**
    -   All references to the moved use cases in `GameManagementUseCases.kt`, `ElapsedTimerManager.kt`, `SudokuGameViewModel.kt`, Koin module (`Module.kt`), and unit tests have been updated to reflect the new package structure.

This reorganization improves the modularity and clarity of the domain layer by grouping related use cases.
