+++ domain/game/src/main/java/com/example/domain/game/Module.kt
@@ -1,25 +1,25 @@
 package com.example.domain.game
 
-import com.example.domain.game.usecases.AutoClearNotesUseCase
-import com.example.domain.game.usecases.ClearActiveGameUseCase
-import com.example.domain.game.usecases.ClearHighlightedRowAndColumnUseCase
-import com.example.domain.game.usecases.FocusOnHintCellsUseCase
-import com.example.domain.game.usecases.GenerateHintLogUseCase
-import com.example.domain.game.usecases.GetBestTimeUseCase
-import com.example.domain.game.usecases.GetElapsedTimeUseCase
-import com.example.domain.game.usecases.GetGameUseCase
-import com.example.domain.game.usecases.HighlightMatchingNumbersUseCase
-import com.example.domain.game.usecases.HighlightRowAndColumnUseCase
-import com.example.domain.game.usecases.InputNumberUseCase
-import com.example.domain.game.usecases.ProvideHintUseCase
-import com.example.domain.game.usecases.RedoOperationUseCase
-import com.example.domain.game.usecases.ResetGameUseCase
-import com.example.domain.game.usecases.RevealHintOnGridUseCase
-import com.example.domain.game.usecases.RevealLastHintLogUseCase
-import com.example.domain.game.usecases.SaveElapsedTimeUseCase
-import com.example.domain.game.usecases.SaveGameUseCase
-import com.example.domain.game.usecases.SelectCellUseCase
-import com.example.domain.game.usecases.UndoOperationUseCase
+import com.example.domain.game.usecases.game.AutoClearNotesUseCase
+import com.example.domain.game.usecases.game.ClearActiveGameUseCase
+import com.example.domain.game.usecases.game.GetGameUseCase
+import com.example.domain.game.usecases.game.ResetGameUseCase
+import com.example.domain.game.usecases.game.SaveGameUseCase
+import com.example.domain.game.usecases.hint.FocusOnHintCellsUseCase
+import com.example.domain.game.usecases.hint.GenerateHintLogUseCase
+import com.example.domain.game.usecases.hint.ProvideHintUseCase
+import com.example.domain.game.usecases.hint.RevealHintOnGridUseCase
+import com.example.domain.game.usecases.hint.RevealLastHintLogUseCase
+import com.example.domain.game.usecases.input.InputNumberUseCase
+import com.example.domain.game.usecases.input.RedoOperationUseCase
+import com.example.domain.game.usecases.input.SelectCellUseCase
+import com.example.domain.game.usecases.input.UndoOperationUseCase
+import com.example.domain.game.usecases.time.GetBestTimeUseCase
+import com.example.domain.game.usecases.time.GetElapsedTimeUseCase
+import com.example.domain.game.usecases.time.SaveElapsedTimeUseCase
+import com.example.domain.game.usecases.visuals.ClearHighlightedRowAndColumnUseCase
+import com.example.domain.game.usecases.visuals.HighlightMatchingNumbersUseCase
+import com.example.domain.game.usecases.visuals.HighlightRowAndColumnUseCase
 import com.example.sudoku.sudokuModule
 import org.koin.dsl.module
