+++ domain/game/src/main/java/com/example/domain/game/usecases/game/GetGameUseCase.kt
@@ -1,0 +1,9 @@
+package com.example.domain.game.usecases.game
+
+import com.example.domain.core.Game
+import com.example.domain.core.GameRepository
+import kotlinx.coroutines.flow.Flow
+
+class GetGameUseCase(private val gameRepository: GameRepository) {
+	operator fun invoke(): Flow<Game> = gameRepository.getGame()
+}
