+++ domain/game/src/main/java/com/example/domain/game/usecases/game/SaveGameUseCase.kt
@@ -1,0 +1,10 @@
+package com.example.domain.game.usecases.game
+
+import com.example.domain.core.Game
+import com.example.domain.core.GameRepository
+
+class SaveGameUseCase(private val gameRepository: GameRepository) {
+	suspend operator fun invoke(game: Game) {
+		gameRepository.saveGame(game)
+	}
+}
