+++ domain/game/src/main/java/com/example/domain/game/usecases/hint/FocusOnHintCellsUseCase.kt
@@ -1,4 +1,4 @@
-package com.example.domain.game.usecases
+package com.example.domain.game.usecases.hint
 
 import com.example.sudoku.model.CellAttributes
 import com.example.sudoku.model.SudokuGrid
