+++ domain/game/src/main/java/com/example/domain/game/usecases/input/SelectCellUseCase.kt
@@ -1,5 +1,8 @@
-package com.example.domain.game.usecases
+package com.example.domain.game.usecases.input
 
+import com.example.domain.game.usecases.visuals.ClearHighlightedRowAndColumnUseCase
+import com.example.domain.game.usecases.visuals.HighlightMatchingNumbersUseCase
+import com.example.domain.game.usecases.visuals.HighlightRowAndColumnUseCase
 import com.example.sudoku.model.CellAttributes
 import com.example.sudoku.model.SudokuGrid
 import com.example.sudoku.model.removeAttribute
