+++ domain/game/src/main/java/com/example/domain/game/usecases/time/GetBestTimeUseCase.kt
@@ -1,4 +1,4 @@
-package com.example.domain.game.usecases
+package com.example.domain.game.usecases.time
 
 import com.example.domain.core.GameDifficulty
 import com.example.domain.core.SudokuGridSize
