+++ domain/game/src/test/java/com/example/domain/game/usecases/SelectCellUseCaseTest.kt
@@ -1,5 +1,9 @@
 package com.example.domain.game.usecases
 
+import com.example.domain.game.usecases.input.SelectCellUseCase
+import com.example.domain.game.usecases.visuals.ClearHighlightedRowAndColumnUseCase
+import com.example.domain.game.usecases.visuals.HighlightMatchingNumbersUseCase
+import com.example.domain.game.usecases.visuals.HighlightRowAndColumnUseCase
 import com.example.sudoku.model.CellAttributes
 import com.example.sudoku.model.SudokuGrid
 import com.example.sudoku.model.addAttribute
